REFACTORED & CHANGES
