+++ example/class_example.dart
