+++ lib/geodesy.dart
@@ -1,3 +1,9 @@
 export 'src/geodesy.dart';
 export 'package:latlong2/latlong.dart' hide pi;
 export 'package:geodesy/src/core/core.dart';
+export 'package:geodesy/src/core/GeodesicMeasurements/polygon_area_by_points.dart';
+export 'package:geodesy/src/core/GeodesicMeasurements/polyline_length_by_multiple_points.dart';
+export 'package:geodesy/src/core/GeodeticPointManipulation/calculate_destination_point.dart';
+export 'package:geodesy/src/core/GeodeticPointManipulation/calculate_points_along_great_circle.dart';
+export 'package:geodesy/src/core/GeodeticPointManipulation/mid_point_between_two_points.dart';
+export 'package:geodesy/src/core/IntersectionAndProjection/geodesic_lines.dart';
