+++ lib/src/core/GeoBounds/get_rectangle_bounds.dart
@@ -1,8 +1,22 @@
-import 'core.dart';
+import '../core.dart';
 
-///
+/// Rectangle Bounds class for calculating the bounding rectangle of a set of LatLng points.
 class RectangleBounds {
-  /// GetRectangleBounds
+  /// Calculates the bounding rectangle (four corner points) that encloses a list
+  /// of LatLng coordinates.
+  ///
+  /// [polygonCoords] - A list of LatLng coordinates to calculate the bounding
+  ///                   rectangle for.
+  ///
+  /// Returns a list of LatLng points representing the corners of the bounding
+  /// rectangle, ordered in the following sequence:
+  ///
+  /// 1. Bottom-left corner
+  /// 2. Bottom-right corner
+  /// 3. Top-right corner
+  /// 4. Top-left corner
+  ///
+  /// The bounding rectangle encompasses all the points provided in the input list.
   static List<LatLng> getRectangleBounds(List<LatLng> polygonCoords) {
     num minLatitude = double.infinity.toDouble();
     num maxLatitude = double.negativeInfinity.toDouble();
