+++ lib/src/core/GeoSpatial/bounding_box.dart
@@ -1,8 +1,17 @@
-import 'core.dart';
+import '../core.dart';
 
-///
+/// A BoundingBox class for calculating bounding boxes and checking if a point is within a bounding box.
 class BoundingBox {
-  /// Calculate  Bounding Box per distance in Kilometers
+  /// Calculates the bounding box (rectangle) around a center point with a specified
+  /// distance in kilometers.
+  ///
+  /// [centerPoint] - The LatLng coordinates of the center point.
+  ///
+  /// [distanceInKm] - The distance in kilometers from the center point to each side
+  ///                   of the bounding box.
+  ///
+  /// Returns a list of two LatLng points representing the top-left and bottom-right
+  /// corners of the bounding box.
   static List<LatLng> calculateBoundingBox(
       LatLng centerPoint, num distanceInKm) {
     // Earth's radius in kilometers
@@ -21,7 +30,16 @@ class BoundingBox {
     return [topLeft, bottomRight];
   }
 
-  /// Check if a given geo point is in the bounding box
+  /// Checks if a given geo point is inside the bounding box defined by the top-left
+  /// and bottom-right corners.
+  ///
+  /// [l] - The LatLng coordinates of the point to check.
+  ///
+  /// [topLeft] - The LatLng coordinates of the top-left corner of the bounding box.
+  ///
+  /// [bottomRight] - The LatLng coordinates of the bottom-right corner of the bounding box.
+  ///
+  /// Returns true if the point is within the bounding box; otherwise, returns false.
   static bool isGeoPointInBoundingBox(
       LatLng l, LatLng topLeft, LatLng bottomRight) {
     return (bottomRight.latitude <= l.latitude &&
