+++ lib/src/core/GeodeticPointManipulation/calculate_destination_point.dart
@@ -1,14 +1,23 @@
 import '../core.dart';
 
-/// Geodetic Point Manipulation - Rhumb Line Destination Formula
-
+/// A Destination Point class for calculating destination points on the Earth's surface
+/// using the Rhumb Line Destination formula.
 class DestinationPoint {
-  /// Calculate the destination point given an initial point
-  /// (latitude and longitude), a bearing (azimuth), and a distance,
-  /// you can use the formula known as the "Direct Rhumb Line" formula or
-  /// "Rhumb Line Destination" formula.
-  /// This formula provides an approximate destination point when traveling
-  ///  along a constant compass direction.
+  /// Calculates the destination point on the Earth's surface given an initial point
+  /// (latitude and longitude), a bearing (azimuth) in degrees, and a distance in
+  /// kilometers. This formula provides an approximate destination point when
+  /// traveling along a constant compass direction.
+  ///
+  /// [initialPoint] - The initial LatLng point representing the starting location.
+  ///
+  /// [bearingDegrees] - The bearing or azimuth angle in degrees, indicating the
+  ///                    compass direction from the initial point.
+  ///
+  /// [distanceKm] - The distance in kilometers to travel along the constant bearing
+  ///                direction from the initial point to the destination point.
+  ///
+  /// Returns a new LatLng object representing the destination point's latitude and
+  /// longitude coordinates.
   static LatLng calculateDestinationPoint(
       LatLng initialPoint, double bearingDegrees, double distanceKm) {
     final double radius = 6371.0; // Earth's radius in kilometers
