+++ lib/src/core/core.dart
@@ -1,16 +1,17 @@
 export 'package:geodesy/src/math/to_radian.dart';
-export 'package:geodesy/src/core/bearing_between_two_geo_points.dart';
-export 'package:geodesy/src/core/bounding_box.dart';
-export 'package:geodesy/src/core/points_in_range.dart';
-export 'package:geodesy/src/core/get_rectangle_bounds.dart';
-export 'package:geodesy/src/core/get_polygon_intersection.dart';
-export 'package:geodesy/src/core/find_polygon_centroid.dart';
+export 'package:geodesy/src/core/GeographicBearing/bearing_between_two_geo_points.dart';
+export 'package:geodesy/src/core/GeoSpatial/bounding_box.dart';
+export 'package:geodesy/src/core/DistanceCalculations/points_in_range.dart';
+export 'package:geodesy/src/core/GeoBounds/get_rectangle_bounds.dart';
 export 'package:geodesy/src/core/DistanceCalculations/geo_points.dart';
-export 'package:geodesy/src/core/vincenty_distance_calculation.dart';
-export 'package:geodesy/src/core/destination_point_by_distance_and_bearing.dart';
-export 'package:geodesy/src/core/cross_track_distance_to.dart';
+export 'package:geodesy/src/core/DistanceCalculations/vincenty_distance_calculation.dart';
+export 'package:geodesy/src/core/GeographicBearing/destination_point_by_distance_and_bearing.dart';
+export 'package:geodesy/src/core/DistanceCalculations/cross_track_distance_to.dart';
 export 'package:geodesy/src/math/to_degrees.dart';
 export 'package:geodesy/src/core/DistanceCalculations/equirectangular_distance.dart';
+export 'package:geodesy/src/core/Polygon/find_polygon_centroid.dart';
+export 'package:geodesy/src/core/Polygon/get_polygon_intersection.dart';
+export 'package:geodesy/src/core/Polygon/polygon_with_hole.dart';
 export 'package:geodesy/src/core/DistanceCalculations/spherical_law_of_cosines_distance.dart';
 export 'package:latlong2/latlong.dart' hide pi;
 export 'dart:math';
